requirements.txt
0x-contract-addresses==3.0.0
0x-contract-artifacts==3.0.0
0x-json-schemas==2.1.0
0x-order-utils==4.0.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.5.2
attrs==23.2.0
base58==2.1.1
bitarray==2.9.2
cbor2==5.6.2
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.0
cloudinary==1.39.1
cytoolz==0.12.3
dacite==1.8.1
Deprecated==1.2.14
Django==4.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.7.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.4.1
gunicorn==20.1.0
hexbytes==0.3.1
idna==3.6
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multiaddr==0.0.9
multidict==6.0.5
mypy_extensions==0.4.4
netaddr==1.2.1
parsimonious==0.8.1
protobuf==3.20.3
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pyee==9.1.1
PyJWT==2.8.0
pymerkle==2.0.2
pyrsistent==0.20.0
python-dotenv==1.0.1
pytz==2022.7.1
pyunormalize==15.1.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rlp==2.0.1
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.2
stringcase==1.2.0
thirdweb-contract-wrappers==2.0.4
thirdweb-eth-account==0.6.6
thirdweb-sdk==3.1.1
toolz==0.12.1
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
varint==1.0.2
web3==5.27.0
websockets==9.1
whitenoise==6.2.0
wrapt==1.16.0
yarl==1.9.4
